build-dev-container: add 1000 user if it does not exist, clone dotfiles

+++ scripts/.local/bin/scripts/build-dev-container.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
-set -e
+set -euo pipefail
 
 DOCKERFILE_PATH=""
+OG_IMAGE_NAME=""
 
 usage()
 {
@@ -39,6 +40,10 @@ parse_arguments()
 
 set_dockerfile_path()
 {
+    if [ -n "${DOCKERFILE_PATH}" ]; then
+        return
+    fi
+
     DOCKERFILES=$(find . -iname "*Dockerfile*")
 NL='
 '
@@ -71,23 +76,77 @@ NL='
     DOCKERFILE_PATH="${DOCKERFILES_ARR[(($NUM-1))]}"
 }
 
+build_og_image()
+{
+    OG_IMAGE_NAME="devcontainer_$(uuidgen)"
+    docker build -t "${OG_IMAGE_NAME}" -f "${DOCKERFILE_PATH}" .
+}
+
+cleanup()
+{
+    if [ -n "${OG_IMAGE_NAME}" ]; then
+        docker image rm "${OG_IMAGE_NAME}"
+        docker system prune
+    fi
+}
+
+set_username()
+{
+    # One weakness here, what if there is a group with id 1000, without a user with id 1000?
+    USERNAME=$(docker run --rm "${OG_IMAGE_NAME}" sh -c 'grep "1000:1000" /etc/passwd | cut -d: -f1')
+}
+
+build_overlay()
+{
+    CONTEXT=$(mktemp -d)
+    DEV_USER="dev"
+
+    docker build -t devcontainer_jonas -f - "${CONTEXT}" <<-EOF
+        FROM $OG_IMAGE_NAME
+
+        # First, add a user with id 1000 if it does not exist
+        RUN set -eux \
+            if [ "$USERNAMEx" = "x" ]; then \
+                if usermod -v > /dev/null 2>&1; then \
+                    # We are prolly in arch-/debian-/fedoraland \
+                    groupadd -g 1000 $DEV_USER; \
+                    useradd -m -u 1000 -g 1000 -s /bin/sh $DEV_USER; \
+                else \
+                    # We are prolly in alpine-with-just-busybox-land \
+                    addgroup -g 1000 $DEV_USER; \
+                    useradd -D -u 1000 -G $DEV_USER $DEV_USER; \
+                fi
+
+        # Then, we add the actual nix environment
+        RUN mkdir /nix && chown -R ${USERNAME:-$DEV_USER}
+
+        ENV USER=${USERNAME:-$DEV_USER}
+        ENV HOME=/home/${USERNAME:-$DEV_USER}
+        USER ${USERNAME:-$DEV_USER}
+        WORKDIR /home/${USERNAME:-$DEV_USER}
+
+        RUN git clone https://github.com/jonifndef/.dotfiles && \
+            ln -s .dotfiles/home-manager/.config/home-manager .config/home-manager
+            ln -s .dotfiles/nix/.config/nix .config/nix
+
+        #RUN curl -L https://nixos.org/nix/install | sh
+
+        #ENV PATH="/home/${USERNAME:-$DEV_USER}/.nix-profile/bin:${PATH}"
+
+        #RUN nix run home-manager/master -- switch --flake .config/home-manager#ubuntu --impure
+EOF
+}
+
 main()
 {
     parse_arguments "$@"
+    set_dockerfile_path
+    build_og_image
+    set_username
+    build_overlay
+    cleanup
 
-    if [ -z "${DOCKERFILE_PATH}" ]; then
-        set_dockerfile_path
-    fi
-
-    OG_IMAGE_NAME="devcontainer_$(uuidgen)"
-    docker build -t "$OG_IMAGE_NAME" -f "${DOCKERFILE_PATH}" .
-
-    # Build overlay here
-
-    docker image rm "${OG_IMAGE_NAME}"
-    docker system prune
-
-    echo "SUCCUESS"
+    echo "SUCCESS"
 }
 
 main "$@"
